:recycle: CategoryItem 컴포넌트 내부 로직 단순화
