--- frontend/src/components/PairRoom/ReferenceCard/CategoryManagementModal/CategoriesEditor/CategoriesEditor.styles.ts
@@ -1,14 +1,0 @@
-import styled from 'styled-components';
-
-export const CategoryList = styled.ul`
-  display: flex;
-  flex-direction: column-reverse;
-  gap: 1rem;
-
-  width: 100%;
-`;
-
-export const CategoryInput = styled.input`
-  width: 100%;
-  border: 1px solid;
-`;
