+++ frontend/src/components/PairRoom/ReferenceCard/CategoryManagementModal/CategoryManagementModal.styles.ts
@@ -16,8 +16,9 @@ export const Header = styled.div`
   align-items: center;
   gap: 1rem;
 
-  color: ${({ theme }) => theme.color.black[80]};
-  font-size: ${({ theme }) => theme.fontSize.lg};
+  color: ${({ theme }) => theme.color.primary[700]};
+  font-size: ${({ theme }) => theme.fontSize.h5};
+  font-weight: ${({ theme }) => theme.fontWeight.medium};
 `;
 
 export const Footer = styled.form`
@@ -27,7 +28,6 @@ export const Footer = styled.form`
   gap: 1rem;
 
   width: 100%;
-  height: 7rem;
 `;
 
 export const AddNewCategoryInput = styled.div`
@@ -39,4 +39,17 @@ export const buttonStyles = css`
   width: 4.4rem;
   height: 4rem;
   border-radius: 0.6rem;
+`;
+
+export const CategoryList = styled.ul`
+  display: flex;
+  flex-direction: column-reverse;
+  gap: 1rem;
+
+  width: 100%;
+`;
+
+export const CategoryInput = styled.input`
+  width: 100%;
+  border: 1px solid;
 `;
