+++ frontend/src/components/PairRoom/ReferenceCard/CategoryManagementModal/CategoryManagementModal.tsx
@@ -1,10 +1,9 @@
-import { FaFilter } from 'react-icons/fa';
 import { LuPlus } from 'react-icons/lu';
 
 import Button from '@/components/common/Button/Button';
 import Input from '@/components/common/Input/Input';
 import { Modal } from '@/components/common/Modal';
-import CategoriesEditor from '@/components/PairRoom/ReferenceCard/CategoryManagementModal/CategoriesEditor/CategoriesEditor';
+import CategoryItem from '@/components/PairRoom/ReferenceCard/CategoryManagementModal/CategoryItem/CategoryItem';
 import { Category } from '@/components/PairRoom/ReferenceCard/ReferenceCard.type';
 
 import useInput from '@/hooks/common/useInput';
@@ -35,6 +34,7 @@ const CategoryManagementModal = ({
   handleSelectCategory,
 }: CategoryManagementModalProps) => {
   const { value, handleChange, resetValue, message, status } = useInput('');
+
   const addCategory = useAddCategory();
 
   const closeCategoryManagementModal = () => {
@@ -44,33 +44,41 @@ const CategoryManagementModal = ({
 
   const handleAddCategorySubmit = (event: React.FormEvent) => {
     event.preventDefault();
+
     if (status === 'ERROR') return;
+
     addCategory.mutateAsync({ category: value, accessCode }).then(() => resetValue());
   };
 
   return (
-    <Modal isOpen={isOpen} close={closeCategoryManagementModal} size="45rem">
+    <Modal isOpen={isOpen} close={closeCategoryManagementModal} size="50rem">
       <Modal.Header>
         <S.Header>
-          <FaFilter />
-          <p>카테고리 선택</p>
+          <p>카테고리 선택하기</p>
         </S.Header>
       </Modal.Header>
       <Modal.CloseButton close={closeCategoryManagementModal} />
       <Modal.Body>
-        <CategoriesEditor
-          categories={categories}
-          selectedCategory={selectedCategory}
-          handleSelectCategory={handleSelectCategory}
-          accessCode={accessCode}
-          closeModal={closeCategoryManagementModal}
-        />
+        <S.CategoryList>
+          {categories.map((category) => (
+            <CategoryItem
+              closeModal={closeModal}
+              accessCode={accessCode}
+              key={category.id}
+              categoryName={category.value}
+              categoryId={category.id}
+              isChecked={category.id === selectedCategory}
+              handleSelectCategory={handleSelectCategory}
+            />
+          ))}
+        </S.CategoryList>
       </Modal.Body>
       <S.Footer onSubmit={handleAddCategorySubmit}>
         <S.AddNewCategoryInput>
           <Input
             value={value}
             placeholder="+ 새로운 카테고리를 입력해주세요."
+            height="4rem"
             onChange={(event) => handleChange(event, validateCategory(event.target.value, isCategoryExist))}
             status={status}
             message={message}
